Refactor/housekeeping (#12)
* refactor: replace magic numbers with named constants and refine comments in bootloader

* feat: add boolean, size type and NULL

* refactor: replace isr hardcoding with isr_table, add ssize_t, refine type casting

* fix: fix bug in vga_print_string

* docs: fix wrong comment

+++ include/video.h
@@ -12,6 +12,6 @@ struct cursor_t {
 };
 
 void clear_vga();
-i32 vga_print_string(const char* buf, const i32 len);
+ssize_t vga_print_string(const char* buf, const ssize_t len);
 
 #endif
